Linux and cellular change only: Pi 3B+ warning removal and CodeChecker fix. (#1059)
A cast is added to avoid a warning in the SPI porting code on Linux when compiling for a Pi 3B+ (which has 32-bit pointers).  While investigating the problem a CodeChecker fix was made in uCellPrivateFileListFirst().

+++ port/platform/linux/src/u_port_spi.c
@@ -320,8 +320,13 @@ int32_t uPortSpiControllerSendReceiveBlock(int32_t handle, const char *pSend,
         errorCodeOrReceiveSize = (int32_t)U_ERROR_COMMON_PLATFORM;
         if (checkOpenDevice(handle)) {
             struct spi_ioc_transfer transf = {0};
-            transf.tx_buf = (__u64)(pOutBuff ? pOutBuff : pSend);
-            transf.rx_buf = (__u64)(pInBuff ? pInBuff : pReceive);
+            // Note: these two pointers, in the spi_ioc_transfer struct,
+            // are just 64 bit unsigned integers, so some care is required
+            // in passing what has to be a pointer to a byte-wise buffer
+            // into them; specifically the uintptr_t is needed on 32-bit
+            // machines (e.g. Pi 3B+) to avoid a compiler warning
+            transf.tx_buf = (__u64)(uintptr_t)(pOutBuff ? pOutBuff : pSend);
+            transf.rx_buf = (__u64)(uintptr_t)(pInBuff ? pInBuff : pReceive);
             transf.len = (unsigned int)len;
             transf.speed_hz = (unsigned int)gSpiCfg[handle].devCfg.frequencyHertz;
             transf.bits_per_word = (unsigned char)(gSpiCfg[handle].devCfg.wordSizeBytes * 8);
